# documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "REST API Drive On BlockChain",
    "description": "REST API usada como suporte a aplicação mmobila do projeto Nomad.\n"
  },
  "host": "https://donomad-api.herokuapp.com/",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "paths": {
    "/farm": {
      "post": {
        "tags": [
          "farm"
        ],
        "summary": "Adicionar Fazenda na Aplicação",
        "description": "",
        "operationId": "addFarm",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Farm object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Farm"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:farms",
              "read:farms"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "farm"
        ],
        "summary": "Update an existing farm",
        "description": "",
        "operationId": "updateFarm",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Farm object that needs to be added to the application",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Farm"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Farm not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:farms",
              "read:farms"
            ]
          }
        ]
      }
    },
...
    "/label": {
      "post": {
        "tags": [
          "label"
        ],
        "summary": "Adicionar Etiqueta na Aplicação",
        "description": "",
        "operationId": "addLabel",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Label"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:farms",
              "read:farms"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "label"
        ],
        "summary": "Atualização de Etiqueta",
        "description": "",
        "operationId": "updateLabel",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Label"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Label not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:label",
              "read:label"
            ]
          }
        ]
      }
    },
...
    "/graintype": {
      "post": {
        "tags": [
          "graintype"
        ],
        "summary": "Add a new grain type to the Application",
        "description": "",
        "operationId": "addGrainType",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Grain Type object that needs to be added to the farm",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GrainType"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:farms",
              "read:farms"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "graintype"
        ],
        "summary": "Update an existing grain type",
        "description": "",
        "operationId": "updateGrainType",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Grain Type object that needs to be added to the farm",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GrainType"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Grain Type not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "api_auth": [
              "write:farms",
              "read:farms"
            ]
          }
        ]
      }
    },
...
    "/production": {
      "post": {
        "tags": [
          "production"
        ],
        "summary": "Production Info from a farm",
        "description": "",
        "operationId": "productionInfo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "production info for a farm",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Production"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Production"
            }
          },
          "400": {
            "description": "Invalid Production Info."
          }
        }
      }
    },
...
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user",
        "description": "This can only be done by the logged in user.",
        "operationId": "createUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/user/signin": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "signin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "The email for login",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login in clear text",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Production": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "farmId": {
          "type": "integer",
          "format": "int64"
        },
        "graintypeId": {
          "$ref": "#/definitions/GrainType"
        },
        "keyuser": {
          "$ref": "#/definitions/User"
        },
        "harvestDate": {
          "type": "string",
          "format": "date-time"
        },
        "farmer": {
          "$ref": "#/definitions/User"
        },
        "altitudeData": {
          "type": "string"
        },
        "goodQuantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "xml": {
        "name": "Production"
      }
    },
    "GrainType": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "graintype": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "usage": {
          "type": "boolean",
          "default": true
        }
      },
      "xml": {
        "name": "GrainType"
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "permissionLevel": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string",
          "description": "User Status",
          "enum": [
            "new",
            "locked",
            "released"
          ]
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "GeoData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "farmid": {
          "type": "integer",
          "format": "int64"
        },
        "altitudedata1": {
          "type": "string"
        },
        "altitudedata2": {
          "type": "string"
        },
        "geodata1": {
          "type": "string"
        },
        "geodata2": {
          "type": "string"
        },
        "gyrodata1": {
          "type": "string"
        },
        "gyrodata2": {
          "type": "string"
        },
        "touchdate": {
          "type": "string",
          "format": "date-time"
        },
        "harvestdate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "xml": {
        "name": "GeoData"
      }
    },
    "Farm": {
      "type": "object",
      "required": [
        "name",
        "photoUrls"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "Fazendo Dois Irmaos"
        },
        "region": {
          "type": "string",
          "example": "Triangulo Mineiro"
        },
        "city": {
          "type": "string",
          "example": "Uberaba"
        },
        "description": {
          "type": "string"
        },
        "usage": {
          "type": "boolean",
          "default": true
        },
        "keyuser": {
          "$ref": "#/definitions/User"
        },
        "geodata": {
          "$ref": "#/definitions/GeoData"
        },
        "certificates": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Farm"
      }
    },
    "Label": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "orderno": {
          "type": "string",
          "example": "21342564475"
        },
        "process": {
          "type": "string",
          "example": "WOLFF"
        },
        "operation": {
          "type": "string",
          "example": "TORRA"
        },
        "production": {
          "$ref": "#/definitions/set01"
        },
        "year": {
          "type": "string"
        },
        "lastsequence": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "comments": {
          "type": "string"
        },
        "usage": {
          "type": "boolean",
          "default": true
        }
      },
      "xml": {
        "name": "Farm"
      }
    },
...
  },
...
}